+++ lib/kitchen/directions/bake_chapter_key_equations.rb
@@ -2,34 +2,38 @@
 
 module Kitchen
   module Directions
-    # Bake directions for eoc key equations
+    # Bake directons for eoc key equations
     #
     module BakeChapterKeyEquations
-      def self.v1(chapter:, metadata_source:, append_to: nil)
-        metadata_elements = metadata_source.children_to_keep.copy
-
-        chapter.key_equations.search('h3').trash
-        key_equations = chapter.key_equations.cut
-
-        return if key_equations.none?
-
-        append_to_element = append_to || chapter
-        title_tag = append_to ? 'h3' : 'h2'
-
-        append_to_element.append(child:
-          <<~HTML
-            <div class="os-eoc os-key-equations-container" data-type="composite-page" data-uuid-key=".key-equations">
-              <#{title_tag} data-type="document-title">
-                <span class="os-text">#{I18n.t(:eoc_key_equations)}</span>
-              </#{title_tag}>
-              <div data-type="metadata" style="display: none;">
-                <h1 data-type="document-title" itemprop="name">#{I18n.t(:eoc_key_equations)}</h1>
-                #{metadata_elements.paste}
-              </div>
-              #{key_equations.paste}
-            </div>
-          HTML
+      def self.v1(chapter:, metadata_source:, append_to: nil, uuid_prefix: '.')
+        V1.new.bake(
+          chapter: chapter,
+          metadata_source: metadata_source,
+          append_to: append_to,
+          uuid_prefix: uuid_prefix
         )
+      end
+
+      class V1
+        renderable
+        def bake(chapter:, metadata_source:, append_to:, uuid_prefix:)
+          @metadata = metadata_source.children_to_keep.copy
+          @klass = 'key-equations'
+          @title = I18n.t(:eoc_key_equations)
+          @uuid_prefix = uuid_prefix
+
+          chapter.key_equations.search('h3').trash
+
+          return if chapter.key_equations.none?
+
+          @content = chapter.key_equations.cut.paste
+
+          append_to_element = append_to || chapter
+          @in_composite_chapter = append_to_element[:'data-type'] == 'composite-chapter'
+
+          append_to_element.append(child: render(file:
+            '../templates/eoc_section_title_template.xhtml.erb'))
+        end
       end
     end
   end