+++ lib/kitchen/directions/bake_chapter_summary.rb
@@ -5,50 +5,57 @@ module Kitchen
     # Bake directions for eoc summary
     #
     module BakeChapterSummary
-      def self.v1(chapter:, metadata_source:)
-        metadata_elements = metadata_source.children_to_keep.copy
-
-        summaries = Clipboard.new
-
-        # TODO: include specific page types somehow without writing it out
-        chapter.non_introduction_pages.each do |page|
-          summary = page.summary
-          summary.first("[data-type='title']")&.trash # get rid of old title if exists
-          summary_title = page.title.copy
-          summary_title.name = 'h3'
-          summary_title.replace_children(with: <<~HTML
-            <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
-            <span class="os-divider"> </span>
-            <span class="os-text" data-type="" itemprop="">#{summary_title.children}</span>
-          HTML
-          )
-
-          summary.prepend(child:
-            <<~HTML
-              <a href="##{page.title.id}">
-                #{summary_title.paste}
-              </a>
-            HTML
-          )
-          summary.cut(to: summaries)
-        end
-
-        return if summaries.none?
-
-        chapter.append(child:
-          <<~HTML
-            <div class="os-eoc os-summary-container" data-type="composite-page" data-uuid-key=".summary">
-              <h2 data-type="document-title">
-                <span class="os-text">#{I18n.t(:eoc_summary_title)}</span>
-              </h2>
-              <div data-type="metadata" style="display: none;">
-                <h1 data-type="document-title" itemprop="name">#{I18n.t(:eoc_summary_title)}</h1>
-                #{metadata_elements.paste}
-              </div>
-              #{summaries.paste}
-            </div>
-          HTML
+      def self.v1(chapter:, metadata_source:, append_to: nil, uuid_prefix: '.')
+        V1.new.bake(
+          chapter: chapter,
+          metadata_source: metadata_source,
+          append_to: append_to,
+          uuid_prefix: uuid_prefix
         )
+      end
+
+      class V1
+        renderable
+        def bake(chapter:, metadata_source:, uuid_prefix:, append_to: nil)
+          @metadata = metadata_source.children_to_keep.copy
+          @klass = 'summary'
+          @title = I18n.t(:eoc_summary_title)
+          @uuid_prefix = uuid_prefix
+
+          summaries = Clipboard.new
+
+          # TODO: include specific page types somehow without writing it out
+          chapter.non_introduction_pages.each do |page|
+            summary = page.summary
+            summary.first("[data-type='title']")&.trash # get rid of old title if exists
+            summary_title = page.title.copy
+            summary_title.name = 'h3'
+            summary_title.replace_children(with: <<~HTML
+              <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
+              <span class="os-divider"> </span>
+              <span class="os-text" data-type="" itemprop="">#{summary_title.children}</span>
+            HTML
+            )
+
+            summary.prepend(child:
+              <<~HTML
+                <a href="##{page.title.id}">
+                  #{summary_title.paste}
+                </a>
+              HTML
+            )
+            summary.cut(to: summaries)
+          end
+
+          return if summaries.none?
+
+          @content = summaries.paste
+          append_to_element = append_to || chapter
+          @in_composite_chapter = append_to_element[:'data-type'] == 'composite-chapter'
+
+          append_to_element.append(child: render(file:
+            '../templates/eoc_section_title_template.xhtml.erb'))
+        end
       end
     end
   end