+++ lib/kitchen/directions/bake_further_research.rb
@@ -5,49 +5,56 @@ module Kitchen
     # Bake directions for further research
     #
     module BakeFurtherResearch
-      def self.v1(chapter:, metadata_source:)
-        metadata_elements = metadata_source.children_to_keep.copy
+      def self.v1(chapter:, metadata_source:, append_to: nil, uuid_prefix: '.')
+        V1.new.bake(
+          chapter: chapter,
+          metadata_source: metadata_source,
+          append_to: append_to,
+          uuid_prefix: uuid_prefix)
+      end
 
-        further_researches = Clipboard.new
+      class V1
+        renderable
+        def bake(chapter:, metadata_source:, append_to:, uuid_prefix:)
+          @metadata = metadata_source.children_to_keep.copy
+          @klass = 'further-research'
+          @title = I18n.t(:eoc_further_research_title)
+          @uuid_prefix = uuid_prefix
 
-        chapter.non_introduction_pages.each do |page|
-          further_research = page.first('.further-research')
-          further_research.first("[data-type='title']")&.trash # get rid of old title if exists
-          further_research_title = page.title.copy
-          further_research_title.name = 'h3'
-          further_research_title.replace_children(with: <<~HTML
-            <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
-            <span class="os-divider"> </span>
-            <span class="os-text" data-type="" itemprop="">#{further_research_title.children}</span>
-          HTML
-          )
+          further_researches = Clipboard.new
 
-          further_research.prepend(child:
-            <<~HTML
-              <a href="##{page.title.id}">
-                #{further_research_title.paste}
-              </a>
+          chapter.non_introduction_pages.each do |page|
+            further_research = page.first('.further-research')
+            further_research.first("[data-type='title']")&.trash # get rid of old title if exists
+            further_research_title = page.title.copy
+            further_research_title.name = 'h3'
+            further_research_title.replace_children(with: <<~HTML
+              <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
+              <span class="os-divider"> </span>
+              <span class="os-text" data-type="" itemprop="">#{further_research_title.children}</span>
             HTML
-          )
-          further_research.cut(to: further_researches)
+            )
+
+            further_research.prepend(child:
+              <<~HTML
+                <a href="##{page.title.id}">
+                  #{further_research_title.paste}
+                </a>
+              HTML
+            )
+            further_research.cut(to: further_researches)
+          end
+
+          return if further_researches.none?
+
+          @content = further_researches.paste
+
+          append_to_element = append_to || chapter
+          @in_composite_chapter = append_to_element[:'data-type'] == 'composite-chapter'
+
+          append_to_element.append(child: render(file:
+            '../templates/eoc_section_title_template.xhtml.erb'))
         end
-
-        return if further_researches.none?
-
-        chapter.append(child:
-          <<~HTML
-            <div class="os-eoc os-further-research-container" data-type="composite-page" data-uuid-key=".further-research">
-              <h2 data-type="document-title">
-                <span class="os-text">#{I18n.t(:eoc_further_research_title)}</span>
-              </h2>
-              <div data-type="metadata" style="display: none;">
-                <h1 data-type="document-title" itemprop="name">#{I18n.t(:eoc_further_research_title)}</h1>
-                #{metadata_elements.paste}
-              </div>
-              #{further_researches.paste}
-            </div>
-          HTML
-        )
       end
     end
   end