+++ lib/kitchen/directions/move_exercises_to_eoc/main.rb
@@ -3,12 +3,24 @@
 module Kitchen
   module Directions
     module MoveExercisesToEOC
-      def self.v1(chapter:, metadata_source:, klass:, append_to: nil)
-        V1.new.bake(chapter: chapter, metadata_source: metadata_source, append_to: append_to, klass: klass)
+      def self.v1(chapter:, metadata_source:, klass:, append_to: nil, uuid_prefix: '.')
+        V1.new.bake(
+          chapter: chapter,
+          metadata_source: metadata_source,
+          append_to: append_to,
+          klass: klass,
+          uuid_prefix: uuid_prefix
+        )
       end
 
-      def self.v2(chapter:, metadata_source:, klass:, append_to: nil)
-        V2.new.bake(chapter: chapter, metadata_source: metadata_source, append_to: append_to, klass: klass)
+      def self.v2(chapter:, metadata_source:, klass:, append_to: nil, uuid_prefix: '.')
+        V2.new.bake(
+          chapter: chapter,
+          metadata_source: metadata_source,
+          append_to: append_to,
+          klass: klass,
+          uuid_prefix: uuid_prefix
+        )
       end
     end
   end