+++ lib/kitchen/directions/move_exercises_to_eoc/v2.rb
@@ -6,10 +6,11 @@ module Kitchen::Directions::MoveExercisesToEOC
   class V2
     renderable
 
-    def bake(chapter:, metadata_source:, klass:, append_to: nil)
+    def bake(chapter:, metadata_source:, klass:, append_to: nil, uuid_prefix: '.')
       @klass = klass
       @metadata = metadata_source.children_to_keep.copy
       @title = I18n.t(:"eoc.#{klass}")
+      @uuid_prefix = uuid_prefix
 
       exercise_clipboard = Kitchen::Clipboard.new
 
@@ -45,10 +46,10 @@ module Kitchen::Directions::MoveExercisesToEOC
       HTML
 
       append_to_element = append_to || chapter
-      @tag = append_to ? 'h3' : 'h2'
-      @metadata_title = append_to ? I18n.t(:eoc_exercises_title) : @title
+      @in_composite_chapter = append_to_element[:'data-type'] == 'composite-chapter'
 
-      append_to_element.append(child: render(file: 'review_exercises.xhtml.erb'))
+      append_to_element.append(child: render(file:
+        '../../templates/eoc_section_title_template.xhtml.erb'))
     end
   end
 end